refactor serde functionality into separate module
