+++ google-apis-common/Cargo.toml
@@ -29,6 +29,3 @@ hyper = "^ 0.14"
 http = "^0.2"
 tokio = "^1.0"
 tower-service = "^0.3.1"
-
-[dev-dependencies]
-serde = { version = "^ 1.0", features = ["derive"] }
